07/05/2021 сделал иконки и отшлифовал сайт

+++ delete.php
@@ -5,10 +5,13 @@ $result_post= $mysql -> query(" SELECT * FROM `manuals`
         WHERE `id` = '$update_ID'
         ");
 $post_manual=$result_post->fetch_assoc();
+if(file_exists($post_manual['file'])){
 unlink($post_manual['file']);
+}
+
 $post_ID_1337= $mysql -> query(" DELETE FROM manuals
 WHERE id='$update_ID'
 ");
 
-header("Location: http://cursach/");
+header("Location: index.php");
 ?>
